make helloworld working. need to parse module names for logging
