+++ concrete/celeryconfig.py
@@ -1,0 +1,6 @@
+# TODO: unification of settings for various libraries and modules
+broker_url = 'pyamqp://'
+result_backend = 'rpc://'
+
+task_serializer = 'json'
+result_serializer = 'json'
