+++ concrete/operators/__init__.py
@@ -1,0 +1,3 @@
+from .aws import AWSOperator
+
+__all__ = ["AWSOperator"]
