+++ concrete/operators/aws.py
@@ -1,0 +1,87 @@
+import os
+import socket
+import time
+from textwrap import dedent
+from uuid import UUID
+
+from ..operator_responses import ProjectDirectory
+
+
+class AWSOperator:
+    """
+    Represents an Operator that takes finalized code, and deploys it to AWS
+    """
+
+    def __init__(self):
+        self.SHARED_VOLUME: str = "/shared"
+        self.DIND_BUILDER_HOST: str = "dind-builder"
+        self.DIND_BUILDER_PORT: int = 5000
+
+    def deploy(self, files: ProjectDirectory, project_uuid: UUID):
+        """
+        Creates and puts a docker image with backend_code + server launch logic into AWS ECR.
+        Launches a task with that docker image.
+        """
+        build_dir_name = f"so_{project_uuid}"
+        build_dir_path = os.path.join(self.SHARED_VOLUME, build_dir_name)
+
+        os.makedirs(build_dir_path, exist_ok=True)
+        dockerfile_content = dedent(
+            f"""
+            FROM python:3.11.9-slim-bookworm
+            WORKDIR /app
+            RUN pip install flask concrete-operators
+            COPY . .
+            ENV OPENAI_API_KEY={os.environ['OPENAI_API_KEY']}
+            ENV OPENAI_TEMPERATURE=0
+            CMD ["flask", "run", "--host=0.0.0.0", "--port=80"]
+            """
+        )
+        start_script = dedent(
+            """
+            #!/bin/sh
+            set -e
+            if ! command -v flask &> /dev/null
+            then
+                echo "Flask is not installed. Installing..."
+                pip install flask
+            fi
+
+            if ! pip show concrete-operators &> /dev/null
+            then
+                echo "concrete-operators is not installed. Installing..."
+                pip install concrete-operators
+            fi
+
+            if [ -z "$OPENAI_API_KEY" ]
+            then
+                echo "Error: OPENAI_API_KEY is not set. Please set it before running this script."
+                exit 1
+            fi
+            flask run --host=0.0.0.0 --port=80
+            """
+        )
+
+        for project_file in files.files:
+            file_path = os.path.join(build_dir_path, project_file.file_name)
+            os.makedirs(os.path.dirname(file_path), exist_ok=True)
+            with open(file_path, "w") as f:
+                f.write(project_file.file_contents)
+
+        with open(os.path.join(build_dir_path, "Dockerfile"), "w") as f:
+            f.write(dockerfile_content)
+        with open(os.path.join(build_dir_path, "start.sh"), "w") as f:
+            f.write(start_script)
+
+        max_retries = 5
+        for _ in range(max_retries):
+            try:
+                with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+                    s.connect((self.DIND_BUILDER_HOST, self.DIND_BUILDER_PORT))
+                    s.sendall(build_dir_name.encode())
+                break
+            except Exception as e:
+                print(e)
+                time.sleep(5)
+
+        return True
