+++ concrete/orchestrator.py
@@ -1,17 +1,15 @@
 from collections.abc import AsyncGenerator
 from textwrap import dedent
-from typing import cast
 from uuid import uuid1
 
 from . import prompts
-from .clients import Client, OpenAIClient
 from .operator_responses import (
     PlannedComponents,
     ProjectDirectory,
     ProjectFile,
     Summary,
 )
-from .operators import AWSOperator, Developer, Executive
+from .operators import AWSOperator, developer, executive
 from .state import ProjectStatus, State
 
 
@@ -31,18 +29,12 @@ class SoftwareProject(StatefulMixin):
         self,
         starting_prompt: str,
         orchestrator: "Orchestrator",
-        exec: Executive,
-        dev: Developer,
-        clients: dict[str, Client],
         aws: AWSOperator | None = None,
         deploy: bool = False,
     ):
         self.state = State(self, orchestrator=orchestrator)
         self.uuid = uuid1()  # suffix is unique based on network id
-        self.clients = clients
         self.starting_prompt = starting_prompt
-        self.exec = exec
-        self.dev = dev
         self.aws = aws
         self.orchestrator = orchestrator
         self.results = None
@@ -53,39 +45,37 @@ class SoftwareProject(StatefulMixin):
         """
         Break down prompt into smaller components and write the code for each individually.
         """
-        self.update(status=ProjectStatus.WORKING, actor=self.exec)
+        self.update(status=ProjectStatus.WORKING)
 
-        components = self.exec.plan_components(self.starting_prompt, response_format=PlannedComponents).components
-        yield Executive.__name__, '\n'.join(components)
+        components = executive.plan_components(self.starting_prompt, response_format=PlannedComponents).components
+        yield executive.__name__, '\n'.join(components)
 
         summary = ""
         all_implementations = []
         for component in components:
             # Use communicative_dehallucination for each component
             async for agent_or_implementation, message in communicative_dehallucination(
-                self.exec,
-                self.dev,
                 summary,
                 component,
                 max_iter=0,
             ):
-                if agent_or_implementation in (Developer.__name__, Executive.__name__):
+                if agent_or_implementation in (developer.__name__, executive.__name__):
                     yield agent_or_implementation, message
                 else:  # last result
                     all_implementations.append(agent_or_implementation)
                     summary = message
 
-        files = self.dev.integrate_components(
+        files = developer.integrate_components(
             components, all_implementations, self.starting_prompt, response_format=ProjectDirectory
         )
 
         if self.deploy:
             if self.aws is None:
                 raise ValueError("Cannot deploy without AWSAgent")
-            cast(AWSOperator, self.aws).deploy(files, self.uuid)
+            self.aws.deploy(files, self.uuid)
 
         self.update(status=ProjectStatus.FINISHED)
-        yield Developer.__name__, str(files)
+        yield developer.__name__, str(files)
 
 
 class Orchestrator:
@@ -102,33 +92,21 @@ class SoftwareOrchestrator(Orchestrator, StatefulMixin):
     def __init__(self):
         self.state = State(self, orchestrator=self)
         self.uuid = uuid1()
-        self.clients = {
-            "openai": OpenAIClient(),
-        }
-        self.operators = {
-            "exec": Executive(self.clients),
-            "dev": Developer(self.clients),
-            "aws": AWSOperator(),
-        }
+        self.aws = AWSOperator()
         self.update(status=ProjectStatus.READY)
 
     def process_new_project(self, starting_prompt: str, deploy: bool = False) -> AsyncGenerator[tuple[str, str], None]:
         self.update(status=ProjectStatus.WORKING)
         current_project = SoftwareProject(
             starting_prompt=starting_prompt.strip() or prompts.HELLO_WORLD_PROMPT,
-            exec=self.operators["exec"],
-            dev=self.operators["dev"],
-            aws=self.operators["aws"],
             orchestrator=self,
-            clients=self.clients,
+            aws=self.aws,
             deploy=deploy,
         )
         return current_project.do_work()
 
 
 async def communicative_dehallucination(
-    executive: Executive,
-    developer: Developer,
     summary: str,
     component: str,
     max_iter: int = 1,
@@ -153,7 +131,7 @@ async def communicative_dehallucination(
         f"""Previous Components summarized:\n{summary}
     Current Component: {component}"""
     )
-    yield Executive.__name__, component
+    yield executive.__name__, component
     # Iterative Q&A process
     q_and_a = []
     for _ in range(max_iter):
@@ -162,12 +140,12 @@ async def communicative_dehallucination(
         if question == "No Question":
             break
 
-        yield Developer.__name__, question.text
+        yield developer.__name__, question.text
 
         answer = executive.answer_question(context, question)
         q_and_a.append((question, answer))
 
-        yield Executive.__name__, answer.text
+        yield executive.__name__, answer.text
 
     if q_and_a:
         context += "\nComponent Clarifications:"
@@ -178,11 +156,11 @@ async def communicative_dehallucination(
     # Developer implements component based on clarified context
     implementation = developer.implement_component(context, response_format=ProjectFile)
 
-    yield Developer.__name__, str(implementation)
+    yield developer.__name__, str(implementation)
 
     # Generate a summary of what has been achieved
     summary = executive.generate_summary(summary, implementation, response_format=Summary)
 
-    yield Executive.__name__, str(summary)
+    yield executive.__name__, str(summary)
 
     yield implementation, str(summary)
